add new option to export transcribed audio to SRT files
a new option has been added to the CLI application that allows users to
export transcribed audio to SRT files.

this feature is useful for creating subtitles for videos or audio
recordings.

to use it, simply run the following command:

parlo --export <srt_file_path>

this will create a new SRT file at the specified path, which will
contain the transcribed audio for the current input file. it's worth
noting that this feature is still under development, but it is already
functional and can be used to create SRT files for most audio
recordings.