change the executable name, add a new export option for plain text format, and remove the "--export" argument
the executable is now named "parlo-cli" instead of only "parlo".

this commit adds a new export option for plain text format, which can be
accessed using the --txt and adds the --srt argument for a replacement of the
old --export option, which is not accessible anymore.

to export to plain text, run the following command:

parlo-cli -a <audio> -m <model> --txt <output_file>

to export to subrip subtitle file, run the following command:

parlo-cli -a <audio> -m <model> --srt <output_file>